+++ .cgtest.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * CGTest v1.16.0 by Balint Toth [TBali]
+ * CGTest v1.17.0 by Balint Toth [TBali]
  * A multi-language offline batch test runner for CodinGame (or other) solo I/O puzzles.
  *
  * configuration file
@@ -23,7 +23,7 @@ return [
     // 'clean' => false,
     // 'create' => '',
     // 'test-case' => 'all',
-    // 'slowThreshold' => 5, // in seconds
+    // 'slowThreshold' => 10, // in seconds
     // 'inputPath' => '.tests/input/',
     // 'inputPattern' => '%p_i%t.txt',
     // 'expectedPath' => '.tests/expected/',
@@ -113,6 +113,9 @@ return [
     //     'cleanPatterns' => [
     //         'fileNameWithPathOrPattern',
     //     ],
+    //     'cleanDirectoryPatterns' => [
+    //         'directoryNameWithPathOrPattern',
+    //     ],
     //     'excludePuzzles' => [
     //         'puzzleName',
     //     ],
@@ -127,28 +130,26 @@ return [
     //         ],
     //     ],
     // ],
-    // == Patterns available in 'buildCommand', 'runCommand', 'cleanPatterns':
+    // == Patterns available in 'buildCommand', 'runCommand', 'cleanPatterns', 'cleanDirectoryPatterns':
     //     %l languageName
     //     %p puzzleName
     //     %o outputPath
     //     %b buildPath
     // == Patterns available in 'buildCommand', 'runCommand':
     //     %s sourceFileName (with path and extension)
-    // == Example (not really needed here, as these are the default settings for c):
-    // 'c' => [
-    //     'sourcePath' => 'c/',
-    //     'sourceExtension' => '.c',
-    //     'codinGameVersion' => 'gcc 11.2.0-20',
-    //     'versionCommand' => 'gcc --version',
-    //     'buildCommand' => 'gcc -std=c17 -o %b%p_%l.exe %s -lm -lpthread',
+    // == Patterns available in 'cleanPatterns', 'cleanDirectoryPatterns':
+    //     %d sourcePath
+    // == Example (not really needed here, as these are the default settings for rust):
+    // 'rust' => [
+    //     'sourcePath' => 'rust/',
+    //     'sourceExtension' => '.rs',
+    //     'codinGameVersion' => 'rustc 1.70.0',
+    //     'versionCommand' => 'rustc --version',
+    //     'buildCommand' => 'rustc -C opt-level=3 --edition 2021 %s -o%b%p_%l.exe',
     //     'runCommand' => '%b%p_%l.exe',
-    //     'note' => '',
-    //     'altVersionCommand' => 'clang --version',
-    //     'altBuildCommand' => 'clang -std=c17 -o %b%p_%l.exe %s',
-    //     'altRunCommand' => '%b%p_%l.exe',
-    //     'altNote' => '',
     //     'cleanPatterns' => [
     //         '%b%p_%l.exe',
+    //         '%b%p_%l.pdb',
     //     ],
     //     'excludePuzzles' => [
     //     ],